Move hooks to Behind GM only toggle
